refactor: menu에서 menuGroup 컨텍스트 분리
